Tacky code generation for Chapter 11
Changed from int/long everywhere to mostly long everywhere.

Changed SemanticAnalyzer to inject a symbol table (facepalm).

+++ src/com/plasstech/lang/c/codegen/AsmStaticVariable.java
@@ -1,6 +1,7 @@
 package com.plasstech.lang.c.codegen;
 
-public record AsmStaticVariable(String name, boolean global, int initialValue)
+/** Represents a static variable in the "Assembly Generation" step. Page 235 */
+public record AsmStaticVariable(String name, boolean global, long initialValue)
     implements AsmTopLevel {
 
   @Override
